added in automatic ip address dectection
the app will automatically try to connect to a websocket on its current ip address at port:8080, note this port is hardcoded to be 8080 on both app and server side.

Now if the app and the server are on the same internet,  the app should run, and no need to recompile the app if you connect to a different wifi network.